Fix check unused arguments false positive bug
One of the check for `unused-argument` looks for unused arguments in
`__new__`. Problem, `__new__` have to match the arguments of `__init__`
even if it does not use them.

This behavior was introduced in the following commit:
https://github.com/pylint-dev/pylint/commit/ebb64fcf6f83eb12e7dd86467054d223922a56b0.
There is no details on why `__new__` has been added to the list.

As it's done for special methods with fixed number of parameters, there
should not be unused arguments check for `__new__`, I removed it from
the checked methods.

Fixes https://github.com/pylint-dev/pylint/issues/3670